Add pg type type info to schema (#592)
Adding `PostgresType` into read_schema output and filling it with info
obtained from `typtype` column of `pg_catalog.pg_type`. Adding a test
with different type types and fixing the already existing ones.

partially fixes #385 

for reference: 
> `typtype char`
typtype is b for a base type, c for a composite type (e.g., a table's
row type), d for a domain, e for an enum type, p for a pseudo-type, r
for a range type, or m for a multirange type. See also typrelid and
typbasetype.

from:
https://www.postgresql.org/docs/current/catalog-pg-type.html#CATALOG-PG-TYPE