Simplify fetching rows with composite primary keys during backfill (#595)
This PR simplifies the way we put together the `WHERE` clause of the
query we use to fetch rows during the backfilling process. We no longer
check the type of the returned identifiers, instead every literal is
quoted as before.

I also added a test to check if we can backfilling works properly with
composite primary keys.

Closes https://github.com/xataio/pgroll/issues/582