Use official `goreleaser/goreleaser-cross` image for cross-platform builds (#604)
Stop using the custom `goreleaser-cross` image that was built as part of
#525 and use an official `gorelaser-cross` image instead. The image is
used as the container in which to build the project for multiple
platforms. The image provides C toolchains for all the platforms
supported by `pgroll`.

The official image now uses `goreleaser` `v2.4.8` which contains the fix
for the bug (https://github.com/goreleaser/goreleaser/pull/5298) that we
were using the custom image to work around.