Remove the ability to rename columns with `alter_column` operations (#603)
Remove the ability to rename columns with `alter_column` operations.

#602 added the `rename_column` operation as the preferred way to rename
columns:

```json
{
  "name": "13_rename_column",
  "operations": [
    {
      "rename_column": {
        "table": "employees",
        "from": "role",
        "to": "job_title"
      }
    }
  ]
}
```

As of this PR, `rename_column` operations are the only way to rename
columns, with the ability to do so using `alter_column` operations
removed.

This is a breaking change to the `pgroll` migrations format.

This is the second (and final) part of #601 

Closes #601