Add support for `alter_table` set `NOT NULL` operations with `rename_table` operations (#606)
Ensure that multi-operation migrations combining `alter_column` `SET NOT
NULL` and `rename_table` operations work as expected.

```json
{
  "name": "06_multi_operation",
  "operations": [
    {
      "rename_table": {
        "from": "items",
        "to": "products"
      }
    },
    {
      "alter_column": {
        "table": "products",
        "column": "name",
        "nullable": false,
        "up": "SELECT CASE WHEN name IS NULL THEN 'anonymous' ELSE name END",
        "down": "name || '_from_down_trigger'"
      }
    }
  ]
}
```

This migration renames a table and then sets a column's `nullability` on
the renamed table.

Previously the migration would fail as the `alter_column` operation was
unaware of the changes made by the preceding operation.

Part of #239