Add support for `alter_table` set `NOT NULL` operations with `rename_column` operations (#607)
Ensure that multi-operation migrations combining `alter_column` `SET NOT
NULL` and `rename_column` operations work as expected.

```json
{
  "name": "06_multi_operation",
  "operations": [
    {
      "rename_column": {
        "table": "items",
        "from": "name",
        "to": "item_name"
      }
    },
    {
      "alter_column": {
        "table": "items",
        "column": "item_name",
        "nullable": false,
        "up": "SELECT CASE WHEN item_name IS NULL THEN 'anonymous' ELSE item_name END",
        "down": "item_name || '_from_down_trigger'"
      }
    }
  ]
}
```

This migration renames a column and then sets the renamed column's
`nullability`.

Previously the migration would fail as the `alter_column` operation was
unaware of the changes made by the preceding operation.

Part of #239