Use official `goreleaser/goreleaser-cross` image
Stop using the custom `goreleaser-cross` image that was built as part of
used as the container in which to build the project for multiple
platforms. The image provides C toolchains for all the platforms
supported by `pgroll`.

The official image now uses `goreleaser` `v2.4.8` which contains the fix
for the bug (https://github.com/goreleaser/goreleaser/pull/5298) that we
were using the custom image to work around.