Add support for foreign key constraints in `create_table` operation (#597)
This PR adds support for `foreign_key` constraints in table constraint
definitions in `create_table`.

It includes all settings that are currently available in PostgreSQL 17.
At the moment, this option is superior to other foreign key definitions
in pgroll because it supports `ON UPDATE` and `MATCH` type settings. As
I mentioned in earlier PRs, once this work is finished, I am unifying
the constraint options in all operations.

Example constraint configuration:
```sql
{
  "name": "provider_fk",
  "type": "foreign_key",
  "columns": [
    "provider_id",
    "product_id"
  ],
  "deferrable": false,
  "references": {
    "table": "provider",
    "columns": [
      "id",
      "product_id"
    ],
    "on_delete": "SET DEFAULT",
    "on_delete_set_columns": ["id"],
    "on_update": "CASCADE",
    "match_type": "SIMPLE"
  }
}
```