Retry creating unique index (#613)
Introduce a retry mechanism for creating unique indexes.
First, create unique index concurrently. Then wait until Postgres is
done with the index creation, periodically checking the view
`pg_stat_progress_create_index`. Once it is done, lookup pg_index to see
if the index is validated or not. If it is, we are good to go. If not,
drop the index and try again.

Since we are running select queries on `pg_stat_progress_create_index`
and `pg_index` and we are expecting a real output to see the progress,
this approach doesn't work with fake db. I have added hardcoded
responses for the fake db scenario, so that we can safely process
migrations to update virtual schema as well.
Not sure if this is a good solution or not. Open for discussions.

Tested manually with high load, while 10m rows being inserted. But I
wonder if there's a way to add a test for the "high load" scenario.