Forbid adding generated columns to existing tables (#637)
This is a temporary limitation we have to introduce to make sure
people are not adding generated columns to their existing tables.
Right now, with the current approach if someone adds a new generated
column to their big table, PostgreSQL has to write the new stored
value or generate the new identity for every row, leading to possible
downtime.