Update `Operation.Start` signature to remove callbacks parameter (#675)
Remove the `...backfill.CallbackFn` variadic parameter from the `Start`
method of the `Operation` interface and all its implementations.

The parameter was passed to all `Start` methods of all operations but
not used. This is likely left over from when backfill was moved out of
individual operations and into a separate step when all DDL was complete
(https://github.com/xataio/pgroll/pull/289).