Support table and column rename operations preceding `create_constraint` operations (#674)
Add support for allowing the table and column on which a
`create_constraint` operation acts to be renamed by preceding operations
in the same migration. For example, ensure that a migration like this
works as expected:

```json
{
  "name": "16_multiple_ops",
  "operations": [
    {
      "rename_table": {
        "from": "items",
        "to": "products"
      }
    },
    {
      "rename_column": {
        "table": "products",
        "from": "name",
        "to": "item_name"
      }
    },
    {
      "create_constraint": {
        "table": "products",
        "type": "check",
        "name": "check_item_name",
        "columns": [ "item_name" ],
        "check": "length(item_name) > 3",
        "up": {
          "item_name": "item_name || '-from-up'"
        },
        "down": {
          "item_name": "item_name || '-from-down'"
        }
      }
    }
  ]
}
```

Here, a three operation migration first renames the `items` table to
`products`, renames the `name` field to `item_name` then adds a `CHECK`
constraint to the `item_name` field on the `products` table.

This PR covers the case of adding `CHECK` constraints only; further
(smaller) PRs may be needed for the other constraint types supported by
`create_constraint`.

Part of #239