Alter sequence owner before dropping the column (#668)
Check if there's any sequence for the column before dropping it. If the
column type is serial, `pg_get_serial_sequence` will return the sequence
name. With this PR, we will be changing the owner column of the seqeunce
to the new column so that we can safely drop the old one.
Fixes: #633