Move backfill options to `start` and `migrate` command (#660)
This PR moves the flags `backfill-batch-delay` and `backfill-batch-size`
to
the commands `start` and `migrate`. So these options no longer show up
in the list
of global flags.

I created a new struct `backfill.Config` to store and pass the options
down
to the migration runner. I also moved the default values for batch delay
and batch size
into the package `backfill`. I renamed `options.go` to `config.go`
because it is a better name given the new contents of the file.

Closes #657