Don't add volatile `DEFAULT` values directly
Test the new column's `DEFAULT` expression for volatility. If the
expression is volatile, add the column without a `DEFAULT` and create a
trigger to backfill the column with the default value. The `DEFAULT`
value will be added on migration completion. If the added column is `NOT
NULL`, a `NOT NULL NOT VALID` constraint is created on the field.

If the `DEFAULT` expression is non-volatile, the column can be added
with a `DEFAULT` directly.

For a volatile default, the column default will be set on operation
completion, along with upgrading any `NOT NULL NOT VALID` constraints on
the column.