Make backfill batch selection exclude rows inserted or updated after backfill start (#652)
Backfill only rows present at backfill start. This is third approach to
solving https://github.com/xataio/pgroll/issues/583. The previous two
are:
* https://github.com/xataio/pgroll/pull/634
* https://github.com/xataio/pgroll/pull/648

This is the most direct approach to solving the problem. At the same
time as the up/down triggers are created to perform a backfill, a
`_pgroll_needs_backfill` column is also created on the table to be
backfilled. The column has a `DEFAULT` of `true`; the constant default
ensures that this extra column can be added quickly without a lengthy
`ACCESS_EXCLUSIVE` lock. The column is removed when the the operation is
rolled back or completed.

The up/down triggers are modified to set `_pgroll_needs_backfill` to
false whenever they update a row.

The backfill itself is updated to select only rows having
`_pgroll_needs_backfill` set to `true` - this ensures that only rows
created before the triggers were installed are updated by the backfill.
The backfill process still needs to *read* every row in the table,
including those inserted/updated after backfill start, but only those
rows created before backfill start will be updated.

The main disadvantage of this approach is that backfill now requires an
extra column to be created on the target table.