Re-read schema in between operations during migration completion (#692)
Re-read the schema in between operations when completing a migration.
This gives each operation `Complete` method the most up-to-date picture
of the schema when invoked.

Stop doing the same thing manually by updating the virtual schema in the
few operation `Complete` methods that were doing so.

Part of https://github.com/xataio/pgroll/issues/643