Fix SQL formatting (#733)
Fix the SQL formatting issue that is causing the `main` build to fail.

[pgFormatter](https://github.com/darold/pgFormatter) released version
[5.6](https://github.com/darold/pgFormatter/releases/tag/v5.6) with some
fix to how typenames are formatted in `ALTER TABLE` statements and this
was picked up by the
[backplane/pgFormatter](https://hub.docker.com/r/backplane/pgformatter/tags)
image which only offers a `latest` tag.

We could build our own `pgFormatter` image to allow us to pin to a
specific version if this happens often; for now just update the SQL
file.