Make `name` field in migration JSON optional (#744)
Make the `name` field in a migration JSON file optional.

If a `name` is not provided, the migration filename (minus the `.json`
suffix) is used as the migration name.

```json
{
  "name": "02_create_another_table",    <-- no longer required
  "operations": [
    {
      "create_table": {
        "name": "products",
        "columns": [
          {
            "name": "id",
            "type": "serial",
            "pk": true
          },
          {
            "name": "name",
            "type": "varchar(255)",
            "unique": true
          },
          {
            "name": "price",
            "type": "decimal(10,2)"
          }
        ]
      }
    }
  ]
}
```

The `--name` flag to `pgroll convert` is also changed to default to
empty (rather than the current timestamp) as part of this change.

Closes https://github.com/xataio/pgroll/issues/715