`sql2pgroll`: Convert `CREATE INDEX IF NOT EXISTS`
Convert `CREATE INDEX IF NOT EXISTS` statements to `create_index`
operations as though the `IF NOT EXISTS` clause was not present.

A `CREATE INDEX` statement like this:

```sql
CREATE INDEX IF NOT EXISTS "name" ON "dogs" USING btree ("name");
```

is now converted to:

```json
{
  "operations": [
    {
      "create_index": {
        "columns": {
          "name": {}
        },
        "method": "btree",
        "name": "name",
        "table": "dogs"
      }
    }
  ]
}
```

rather than a raw SQL migration.

See the discussion on https://github.com/xataio/pgroll/issues/719 for
the rationale behind this change.