Dont require database access to `pgroll latest --local` (#760)
Ensure that the `latest` command does not require database access when
the `--local` flag is provided.

In this case the command should act only locally, with the migrations on
disk and not require a connection to the database.

The reason it used to require a DB connection was to support the
`--with-schema` flag which took the schema name from an instantiated
`Roll` instance. Change this so that the schema name is instead taken
from the command line `--schema` flag (or the `PGROLL_SCHEMA` env var)
directly instead.

Fixes https://github.com/xataio/pgroll/issues/726