Use `sigs.k8s.io/yaml` to parse YAML migration files (#771)
This PR replaces the unmaintained `gopkg.in/yaml.v3` repository with
`sigs.k8s.io/yaml`. The problem with the previous dependency is that we
cannot not support values that can be null or empty strings. I
implemented a custom nullable string with the appropriate marshaling and
unmarshaling functions. However, `UnmarshalYAML` is not called on null
values. So we cannot not support unsetting comments and default values
in YAML migrations with the current library.

With this new dependency the problem I mentioned above is not an issue.
Also, we no longer need `yaml` tags and the `UnmarshalYAML` function for
`migrations.Operation`. `UnmarshalJSON` takes care of reading the YAML
files.