Add fallback method for backfilling records (#779)
Previously, `pgroll` could not backfill tables that did not have a
primary key or a unique column. From now on, `pgroll` can fallback onto
the internal column `_pgroll_needs_backfill` to find all columns that
has to be backfilled.

I also moved `CNeedsBackfillColumn` constant to the package `backfill`
because it makes more sense to have it there.

Required by #698