Pass all column attributes to new temporary column when adding a new one to the table (#788)
The following migration was failing due to several reasons:

```json
{
  "operations": [
    {
      "create_table": {
        "name": "users",
        "columns": [
          {
            "name": "id",
            "type": "uuid",
            "pk": true
          }
        ]
      }
    },
    {
      "add_column": {
        "table": "users",
        "column": {
          "name": "name",
          "type": "text",
          "nullable": true
        }
      }
    },
    {
      "create_constraint": {
        "table": "users",
        "columns": [
          "name"
        ],
        "type": "unique",
        "name": "my_custom_name",
        "up": {
          "name": "name"
        },
        "down": {
          "name": "name"
        }
      }
    }
  ]
}
```

This PR fixes the first problem with the migration. The first problem is
when `pgroll` adds the new column with the temporary name to the table
schema it lacks all column attributes. Thus, if a later operation tries
to duplicate this duplicated column it fails because we do not know the
type of the column, the default value, etc. The SQL statement generated
has a syntax error because of the lack of information:

```
Failed to start migration: unable to execute start operation: failed to duplicate column: pq: syntax error at or near ","
```