Make `UnappliedMigrations` method return raw migs
Make `UnappliedMigrations` read migrations from the local filesystem as
raw migrations instead of deserializing them. This allows unapplied
migrations in directories containing migrations in older file formats to
be found, without causing an error.

Change the return type to `RawMigration` to avoid having to parse the
migrations before returning them. It is the caller's responsibility to
parse the migrations, if required.