Add baseline tests for missing migrations
Ensure that a baseline migration is not considered a missing migration.

Baseline migrations are not real migrations, and so they should never be
considered as missing on the local filesystem.

Migrations that were created before a baseline should also not be
considered missing from the local filesystem as a baseline migration is
considered to subsume all migrations that were created before it.