Make `SchemaHistory` method not return baselines
Baseline migrations are not part of the schema history as they represent
migrations that were not applied by `pgroll` so they should not be
included in the schema history.

Migrations that were applied before the most recent baseline migration
are also not part of the schema history. This is because the baseline
subsumes all migrations that were applied before it.