Make `pgroll pull` and `pgroll migrate` command able to handle incompatible migration file formats (#812)
Ensure that the `pgroll pull` and `pgroll migrate` commands are able to
operate safely when either:

* The local migration directory contains migrations that can not be
deserialized by the current version of `pgroll`.
* The migration history in the target database contains migrations that
can not be deserialized by the current version of `pgroll`.

This ensures that these commands work in the presence of older migration
formats, either locally or in the remote schema history.

This is done by adding a new `RawMigration` type to the `migrations`
package, which does not deserialize the `Operations` in the migration
but leaves them as a `json.RawMessage`.

The following methods in the `roll` and `state` packages are then
updated to work with and return `RawMigrations`:

* `SchemaHistory` - used to return the schema history from the target
database.
* `MissingMigrations` - used to find remote migrations that don't exist
in the local history
* `UnappliedMigrations` - used to find local migrations that don't exist
in the remote.

By updating these methods and the `pull` and `migrate` commands that use
them, these commands can work safely even with local or remote
migrations that can't be deserialized by the current version of
`pgroll`.

Tests are added for both the `MissingMigrations` and
`UnappliedMigrations` methods to ensure that they work with
un-deserializable migrations.

Part of https://github.com/xataio/pgroll/issues/770