Add new guide to documentation: Use pgroll with ORMs (#816)
This PR adds a new guide to pgroll's documentation. This is a rehashed
version of the release blog post for `convert`
command. It is adjusted a little, so it stands on its on as a guide.

The examples are in JSON because the output format of the `convert`
command is JSON.