Add new command: `create` to add new migrations (#815)
This PR adds a new command named `create`. It lets users create new
migration files. It has three flags: `--empty` if the user wants to
create an empty file and `--name` if the user wants to pass a name
instead of being prompted. Similarly to `pull` command it also has a
flag called `--json` to output the migration file in json format.

I am adding documentation in a follow-up PR.