Remove redundant `pgroll.internal` parameter
The `pgroll.internal` custom param never worked, because it was
set outside of a transaction using `SET LOCAL`, so it had no effect
other than spitting out a warning to the Postgres logs.

The `pgroll.internal` parameter was meant to be used to ensure that
DDL done by `pgroll` operations was not captured as an `inferred`
migration. However, the `raw_migration` function also uses the
`is_active_migration_period` function for this same purpose, so the
`pgroll.internal` parameter was redundant even if it had worked.