Refactoring of writing migrations to file (#821)
This PR adds a new type `MigrationWriter` to provied a unified way to
write migrations
to file either in YAML or in JSON format. It is used in 3 commands:
`convert`, `create` and `pull`.

Also with this change I added a new flag to `convert` called `json`.
Previously, the command serialized migrations
into JSON format. From now on, it follows the same pattern like `create`
and `pull`. It writes migrations in YAML format unless the user sets
`--json` flag.