[baseline] Make `start` and `migrate` fail on non-empty schema without migration histories (#835)
Make the `pgroll start` and `pgroll migrate` commands fail when run
against a non-empty schema without a migration history. The error
message suggests running `pgroll baseline` before re-running `start` or
`migrate`.

Running an initial migration against a non-empty schema is not
recommended becauses it can result in a migration that can't be rolled
back if the migration references pre-existing objects in the schema.

The recommended workflow is:

1. Create the initial baseline migration
```
$ pgroll baseline 00_initial_baseline migrations/
```

2. Run `migrate` as usual

```
$ pgroll migrate migrations/ --complete
```

This ensures that the full schema is brought into `pgroll` before
running the first migration.

Part of https://github.com/xataio/pgroll/issues/364