New `DBAction`: `CreateFKConstraintAction` (#843)
This PR adds a new `DBAction` for creating foreign key constraints. It
is used in `op_set_fk` and `op_create_constraint`.

This refactoring has uncovered the lack of support of deferrable
constraints on the table level. In a follow-up PR I will add
the two missing options `deferrable` and `initially_deferred` to the
`create_constraint`.

Related to #742