Remove support for the `name` field in migrations (#852)
Remove support for specifying a `name` field in migration files:

```yaml
name: 01_some_migration # <-- no longer supported
operations:
  - create_table:
      name: products
      columns:
        - name: id
          type: serial
          pk: true
        - name: name
          type: varchar(255)
```

The migration filename is the source of truth for the migration name.
The `name` field was made optional and its use discouraged in the
documentation in #744. This PR removes support for it entirely.

⚠️ **Warning**

**This is a breaking change:** migration files that set the `name` field
will not be runnable by `pgroll` until the `name` field is removed.

Because of https://github.com/xataio/pgroll/pull/812 `pgroll pull` and
`pgroll migrate` will still work with migration histories that contain
migrations with `name` fields.

The easiest way to remove the `name` fields from all migrations in the
history is to `pgroll pull` the migration history; this will serialize
all migrations in the remote history, removing the any `name` fields.

---

Closes #715 and #850