Add new subcommand to validate migration files (#856)
This PR adds a new subcommand to validate migration files called
`validate`.

```
Usage:
  pgroll validate <file> [flags]

Examples:
validate migrations/03_my_migration.yaml

Flags:
  -h, --help   help for validate
```

It is first step towards providing a way to validate migration files
before running them. At the moment it can validate a single migration
file against the latest schema in the database. This can be added to
your local workflow, or a linter to make sure that the newly added
migration file is correct.

Future improvements can be
* validating multiple migrations
* validating all unapplied migrations in a folder
* verifying `up` migrations on a subset of the table
https://github.com/xataio/pgroll/issues/169
* run `jsonschema.Validate` during validate
https://github.com/xataio/pgroll/issues/673

The PR contains some minor changes to the migration phase `Start`. Now
migration validation is decoupled from it. The previous implementation
of `Start` started a new migration before validating the migration it
got as a parameter. So this change does not interfere with the UX of
pgroll.

Closes #370

---------

Co-authored-by: Andrew Farries <andyrb@gmail.com>