Set `pgroll.internal` on `pgroll` connections
Set a session-level setting `pgroll.internal` to `TRUE` on connections
used by `pgroll` (`pgroll` uses two connections, one to execute
migrations and one for its internal state).

Use the `pgroll.internal` setting to prevent schema changes made by
`pgroll` itself being captured by the raw_migration() event trigger
function.

DDL changes made in schema `s` are already ignored by `raw_migration()`
if there is an active migration period in progress for schema `s`. The
change in this commit ensures that *all* DDL changes in *all* schema at
*all* times made by `pgroll` are ignored by `raw_migration()`, including
version schema creation and the creation of views within those schema.

This reverts PR #823 which removed a `pgroll.internal` setting that was
set using `SET LOCAL` outside a transaction block. That PR assumeed that
setting `pgroll.internal` with `SET LOCAL` outside a transaction had no
effect, but it did have an effect (the setting was set to an empty
string, which was sufficient given how it was used in`raw_migration()`).
This PR uses the same setting but sets it with `SET` (instead of `SET
LOCAL`) to make it more obvious what is happening.